fix(mastery): 修复排班过滤和无 BOX 兼容并补全预检回归
