+++ arknights_mower/tests/mastery_db_tests.py
@@ -5,6 +5,7 @@ import tempfile
 import unittest
 from unittest.mock import patch
 
+from arknights_mower.tests.mastery_plan_helpers import stub_support_planner
 from arknights_mower.utils.mastery_db import (
     add_plan_checked,
     delete_plan,
@@ -30,12 +31,7 @@ from arknights_mower.utils.mastery_recommendation import get_current_mastery_lev
 
 class TestMasteryDb(unittest.TestCase):
     def setUp(self):
-        planner = patch(
-            "arknights_mower.utils.mastery_support.plan_supports",
-            return_value={"version": 1, "stages": []},
-        )
-        planner.start()
-        self.addCleanup(planner.stop)
+        stub_support_planner(self)
         self.tmp = tempfile.NamedTemporaryFile(suffix=".db", delete=False)
         self.db_path = self.tmp.name
         self.tmp.close()
