+++ doc/mastery-supports.md
@@ -1,12 +1,14 @@
 # 每个计划的专精协助方案
 
-新增计划时，按已同步的森空岛干员练度和当前资源，为尚未完成的专精阶段生成协助方案，并与计划原子保存。计划列表中的「协助方案」可以查看、修改各级初始协助者和后段接班者。开始执行前可以修改全部阶段；训练中允许修改后续阶段，已经开始的阶段锁定。保存时对计划状态、原方案和执行快照做并发校验。
+选择自动协助且有 BOX 时，按已同步的森空岛干员练度和当前资源，为尚未完成的专精阶段生成协助方案，并与计划原子保存。计划列表中的「协助方案」可以查看、修改各级初始协助者和后段接班者。开始执行前可以修改全部阶段；训练中允许修改后续阶段，已经开始的阶段锁定。保存时对计划状态、原方案和执行快照做并发校验。
+
+未同步 BOX（或文件损坏、无干员数据）时，新增计划自动使用原职业路线并提示确认协助者配置，不要求使用森空岛。手动添加任务也可显式选择「使用职业路线」，跳过自动协助规划。已有 BOX 时，未拥有/未精二、资源规则缺失和无可用协助者等校验错误仍会阻止自动建计划，不静默降级。职业路线计划没有逐计划协助方案，直接在通用专精路线预览中设置教官。
 
 ## 排班约束
 
 - 主排班与全部备用排班中的主力、replacement 都计入检查，不判断备用排班是否生效。
 - 出现在任何非训练室设施（包括宿舍、活动室）的干员均从协助者名单中排除；训练室栏不排除。Free、Current 和空值不代表干员。
-- 被训练干员出现在非训练室排班时，添加计划仅弹出警告，仍允许添加、编辑协助方案和执行训练。原有「空闲／非空闲」筛选保留，排班占用仍计为非空闲，不再作为被训练干员的准入限制。
+- 被训练干员出现在非训练室排班时，添加计划仅弹出警告，仍允许添加、编辑协助方案和执行训练。原有「空闲／非空闲」筛选保留，所有设施（含训练室）的主/备用排班主力与替换占用仍计为非空闲，不再作为被训练干员的准入限制。
 - 阿斯卡纶、烛煌或斩业星熊出现在上述任意排班的 **central 主力或 replacement** 中，就自动启用中枢 +5%。多名同时出现仍为 +5%；出现在其他房间不会触发。页面开关只展示推导结果。
 - 自选协助者保存时检查拥有情况、不能自己训练自己和排班冲突，允许选择无专精加速技能的干员。仅实际改过的阶段标记为自选。
 
@@ -42,7 +44,7 @@
 
 ## 执行与换人
 
-1. 开训前回到训练室主页面，核对阶段、槽位、排班约束和上一阶段减半记录，保存本级执行方案。此时不改协助位。
+1. 技能选择页（219）读出档位后返回，等待场景确认为训练室主页面（217），再复用原有 `_read_slots_checked` 读取槽位；过渡场景先等待。核对阶段、槽位、排班约束和上一阶段减半记录，保存本级执行方案。此时不改协助位。
 2. 若上一阶段已获得下次训练减半效果，保留原协助者至本级训练开始、减半生效，再安排本级初始协助者。安排后确认实际槽位，重读倒计时，以截图安排后续任务。
 3. 后段减半换人按实际倒计时和当前速度计算，至少留 5 小时加缓冲（缓冲最少 1 分钟）。首选因排班等原因不可用时，按可用减半教官重新计算接班时机；换人后再次核对槽位和实际倒计时。
 4. 剩余时间不满足减半条件时保留原人；槽位确认或换人失败时停止本级自动换人并告警。通知按计划和等级去重，删除计划时清理对应记录。
@@ -50,20 +52,26 @@
 
 本版本不包含新增的专精心情识别、心情消耗计算、开训前心情检查或按心情替补与告警；协助者能否持续工作需用户自行保证。这部分后续单独实现。
 
-`enable_mastery` 与现有训练室保护规则继续生效。`assistant_follows_schedule=True` 时不执行自动换协助者。旧计划没有协助方案时仍用原职业路线；重新添加计划可生成新方案。资源包缺少 training v1 时明确提示更新资源，不静默使用不完整的规则。
+`enable_mastery` 与现有训练室保护规则继续生效。`assistant_follows_schedule=True` 时不执行自动换协助者。旧计划与职业路线计划仍用原职业路线。自动协助模式有 BOX 但资源包缺少 training v1 时明确提示更新资源，不静默使用不完整规则；显式职业路线模式不要求训练规则。
+
+## 模块与持久化
+
+`mastery_support` 保留规划/预览入口，数据与候选过滤在 `mastery_support_data`，路线求解在 `mastery_optimizer`，方案编辑在 `mastery_support_edits`；共享输入、阶段参数和 JSON 编解码集中在 `mastery_support_types`。执行分为 `mastery_support_runtime`（预检与恢复）、`mastery_support_state`（快照与任务）及 `mastery_support_swap`（校验与换人）。
+
+`GET/PATCH /mastery-plan/supports` 读取/修改逐计划方案，DB 的 `support_plan` 与 `support_runtime` 分离，使用原状态及快照比较防止编辑覆盖正在执行的阶段。`support_swap` 是第九类去重通知，key 为 `{plan_id}:{level}`；删除计划时一并清理。两种建计划载荷均校验技能序号为整数 0/1/2。
 
 ## 性能与验证边界
 
 资源规则由现有资源缓存读取；练度文件按路径、修改时间和大小缓存，更新后自动失效。页面排班范围使用 computed；候选名单只在打开协助方案时读取。不增加选人页扫描，不在调度主循环重新求解。
 
-单元测试覆盖资源规则、路线求解、排班约束、方案保存、阶段校验和换人任务。完整的实机训练流程及不同客户端画质仍需后续验证。
+单元测试覆盖资源规则、路线求解、排班约束、方案保存、阶段校验和换人任务。开训流程测试调用真实预检与槽位读取函数，设备边界模拟技能页返回、UNKNOWN/CONNECTING 过渡及读取失败。完整的实机训练流程及不同客户端画质仍需后续验证。
 
 资源生成器和该功能代码发布后，MowerResource 需要重新出包。此功能不向计算网站发起运行时请求。
 
 验证命令：
 
 ```sh
-.venv/bin/python -m pytest arknights_mower/tests/mastery_support_tests.py -q
-.venv/bin/python -m pytest arknights_mower/tests/*.py -q
+.venv/bin/python -m pytest arknights_mower/tests/mastery_support_*tests.py -q
+.venv/bin/python -m pytest -q  # 自动发现应用与 scripts 测试，兼容 unittest/pytest
 cd ui && npm test && npm run build
 ```
