+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+testpaths = arknights_mower/tests scripts/tests
+python_files = *_tests.py test_*.py
